broke up line module
