+++ autofit/tools/edenise/structure/__init__.py
@@ -1,4 +1,5 @@
 from .file import File
-from .import_ import Import, LineItem
+from .import_ import Import
+from .line import LineItem
 from .item import Item
 from .package import Package
