+++ autofit/tools/edenise/structure/import_.py
@@ -1,122 +1,8 @@
 import ast
-from copy import copy
 from typing import Optional
 
-from astunparse import unparse
-
-from autofit.tools.edenise.structure.item import Item
-
-
-class LineItem(Item):
-    def __init__(
-            self,
-            ast_item: ast.stmt,
-            parent: Optional[Item] = None
-    ):
-        """
-        A line in a file.
-
-        Special consideration is given to lines containing imports.
-
-        When parentheses are not balanced a line can span multiple lines in a file.
-
-        Parameters
-        ----------
-        ast_item
-            The string on the line
-        parent
-            The file containing the line
-        """
-        self.ast_item = ast_item
-        super().__init__(
-            parent=parent
-        )
-
-    def converted(self):
-        return copy(
-            self.ast_item
-        )
-
-    @staticmethod
-    def parse_fragment(
-            string,
-            parent=None
-    ):
-        return LineItem(
-            ast.parse(
-                string
-            ).body[0],
-            parent=parent
-        )
-
-    @property
-    def target_string(self):
-        return unparse(
-            self.converted()
-        )
-
-    def __new__(cls, ast_item, parent):
-        if isinstance(
-                ast_item,
-                ast.Import
-        ):
-            return object.__new__(Import)
-        if isinstance(
-                ast_item,
-                ast.ImportFrom
-        ):
-            return object.__new__(ImportFrom)
-        if isinstance(
-                ast_item,
-                ast.FunctionDef
-        ):
-            return object.__new__(Function)
-        if isinstance(
-            ast_item,
-            ast.Assign
-        ):
-            return object.__new__(Assignment)
-        return object.__new__(LineItem)
-
-
-class Assignment(LineItem):
-    def converted(self):
-        converted = super().converted()
-        aliases = self.parent.aliases
-
-        def strip_ids(
-                obj
-        ):
-            if isinstance(
-                obj, ast.Attribute
-            ) and obj.value.id in aliases:
-                return ast.Name(
-                    id=obj.attr,
-                    col_offset=obj.col_offset,
-                    ctx=obj.ctx,
-                    lineno=obj.lineno
-                )
-            if hasattr(obj, "__dict__"):
-                obj.__dict__ = strip_ids(
-                    obj.__dict__
-                )
-            if isinstance(obj, dict):
-                return {
-                    key: strip_ids(
-                        value
-                    )
-                    for key, value
-                    in obj.items()
-                }
-            if isinstance(obj, list):
-                return list(map(
-                    strip_ids,
-                    obj
-                ))
-            return obj
-
-        strip_ids(converted)
-        return converted
+from .line import LineItem
+from .item import Item
 
 
 class Import(LineItem):
@@ -246,10 +132,3 @@ class ImportFrom(Import):
         return path
 
 
-class Function(LineItem):
-    def converted(self):
-        converted = super().converted()
-        for arg in converted.args.args:
-            arg.annotation = None
-        converted.returns = None
-        return converted
